module5-task1 Update classes with inheritance
